[fix] condense Network tab stats header and improve filter a11y on short-height windows (#257)
The Network tab's stats card (~150dp two-row layout) left almost no
room for the request list on short-height windows, same class of
issue as the Debug Panel filter bar. On compact height it now
condenses to a single pinned line (download/upload/requests/errors/
duration) instead of folding into the scrollable list, since error
count and totals are ambient signal worth keeping glanceable while
scrolling requests, not a control users tuck away. Non-compact stays
unchanged.

Also address PR feedback: the Logcat filter dropdown's trigger button
now announces the currently selected level for screen readers instead
of a static description.
